requirements.txt
Faker==2.0.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.14
marshmallow-sqlalchemy==0.17.0
mysql-connector-python==8.0.11
pytest==5.0.1
voluptuous==0.11.5
